fix: rename packagename to confsched
