+++ core/testing/src/main/java/io/github/droidkaigi/confsched/testing/robot/ProfileCardScreenRobot.kt
@@ -3,10 +3,10 @@ package io.github.droidkaigi.confsched.testing.robot
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.hasTestTag
 import io.github.droidkaigi.confsched.designsystem.theme.KaigiTheme
+import io.github.droidkaigi.confsched.profilecard.ProfileCardScreen
+import io.github.droidkaigi.confsched.profilecard.ProfileCardTestTag
 import io.github.droidkaigi.confsched.testing.DefaultScreenRobot
 import io.github.droidkaigi.confsched.testing.ScreenRobot
-import io.github.droidkaigi.confshed.profilecard.ProfileCardScreen
-import io.github.droidkaigi.confshed.profilecard.ProfileCardTestTag
 import javax.inject.Inject
 
 class ProfileCardScreenRobot @Inject constructor(
